requirements.txt
---

# 📦 requirements.txt

```txt
fastapi
uvicorn
paho-mqtt
pandas
numpy
scikit-learn
joblib
streamlit
sqlite-utils